Added zombie characters
Exported zombie characters from Maya as fbx files along with the basic animations and added them to the unity project. Tested animations in unity by creating an animator controller and playing them in the editor. There were some issues with the animation retargeting. Fixed the animations in maya and exported them as fbx again. Added in unity and tested the animations again. Made texture variants and added textures to the unity project.