Menus
Added UIManager.cs script that is used to manage all the game menus and UI-related jobs. Currently added 3 menus, namely: Game launch/Main Menu, Gameplay, and pause menu. UIManager has a function "ButtonPressed" that listens to all the UI button click events. It takes a string as a parameter that is used to differentiate the functionality of a button. The flow of the menus as as follows:
Upon starting/launching the game, the very first screen visible to the user will be the main menus screen. It has 1 play button. On clicking the play button the Gameplay screen gets enabled, where the user will see his character and zombies. The top right cover contains the distance tab, players current distance travelled will appear there. On the top left corner of the gameplay screen is a pause button. On clicking the pause button, pause screen appears that contains two buttons: The resume button that takes player back to the gameplay screen and the end button that takes player to the main menu screen and quit the game.