Audios and Preferences
Added AudiosManager.cs, a script that is used to play game background music and sound effects with gameplay. It is responsible to manage audio volume and playing sound effects. AudioData data type is made to act as a placeholder for audio clips. AudioData has 2 fields "Audio name" and "Clip". AudiosManager has an array of AudioData and sound clips can be played using this array.
Game data will be saved using PlayerPrefs and a wrapper class for this purpose is created named as PreferenceHandler,cs All the game data that requires to be saved for later purposes is accessed through it. Currently it stores the Music and Sounds volume and best distance travelled by player.
New screen "Settings" is added. Player can set the game music and sounds volume in this menu using sliders.