Bullets
Integrated bullets system. Spawning bullets with the desired firerate according to the gun player has. We can set the firerate in Utilities.cs with 2 constant floats named as PistolFireRate, SMGFireRate for pistol, and smg accordingly. The firerate is the the delay between each bullet fired if the user press and holds Up/W key. A bullet prefab is created in Prefabs/Weapons folder, this prefab gets spawned depending upon the input and firerate. The speed of the bullet can be set in the inspector view of the bullet prefab. The prefab hasCollisionEnter fuction that listens to the bullet hit. If a zombie comes in collision with the bullet, it destroys the bullet and play's the kill animation of the zombie.