Weapons Spawner
Implemented WeaponsSpawner.cs, a script that handles the spawning of weapons. It spawns the SpawnedWeaponsHandler.cs, the script responsible for randomly appearing weapons in the scene. The SpawnedWeaponHandler listens player's collision and then sets the gun (pistol, SMG) in PlayerController accordingly. Both pistol and SMG bullet counts can be set in Utilities.cs script. The weapons spawn time increases in the form of multiples of the value set in inspector of WeaponsSpawner. Currenlty it spawns in following interval of seconds, (2,4,6,8,10,...)