Adding sounds and effects and issues fixes
Added bullets, zombie kill, attack, the player hit, death, pick up guns sounds and effects.
Added skybox in the scene and fog and lighting adjustments.
Added player's running with gun animation and transition in animator.
Fixes:
The animation transition fixes the previous state getting stuck when the player gets a weapon. Issue was with the animation transition state condition, which has been set to the greater than value, whereas it was required to be set as equal to.
The controls issue is fixed, and the left and right strafing value is clamped between the values along with the path. Previously, the player goes out of the path even off-screen if it keeps on pressing the left or right key.
Zombie movement through the game pause screen is fixed. Zombie keeps on moving behind the pause screen, while the player was stopped before. Now only the animation keeps playing.
Fixed scene lighting issues. Upon changing the skybox scene became extra dark.
Fixed zombie attacking the player issues, before the zombie's attack animation gets player way later even when the player passes through it. The collider on the zombie is moved further forward to start attacking before.
The zombie hit-by-bullet issue is fixed. By the solution of the previous issue, now the bullet also hits the zombie way before, as the collider was shifter forward. Fixed this issue by using the other non-trigger collider and using the OnCollisionEnter function instead of OnTriggerEnter.
Fixed the distance counter error, it keeps on counting the distance till the player's death animation gets played, even the player was not moving at all.